fix/feat: @Modifying long→int + Spring Actuator + 구조화 JSON 로깅 + MDC 추적
* fix/feat: @Modifying long→int 수정 + Spring Actuator 메트릭 노출 (admin 전용)

* feat: 구조화 JSON 로깅 (ECS 형식) + 요청 MDC 추적 (X-Request-ID)

---------

+++ jbDataHub/src/main/java/com/jb/datahub/auth/RefreshTokenCleanupScheduler.java
@@ -3,7 +3,6 @@ package com.jb.datahub.auth;
 import com.jb.datahub.auth.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,7 +18,7 @@ public class RefreshTokenCleanupScheduler {
     @Scheduled(cron = "0 0 3 * * *")
     @Transactional
     public void cleanupExpiredTokens() {
-        long deleted = repository.deleteExpiredAndUsed();
+        int deleted = repository.deleteExpiredAndUsed();
         if (deleted > 0) {
             log.info("Cleaned up {} expired/used refresh tokens", deleted);
         }
